refactored into schema and resolver
