+++ server.js
@@ -87,7 +87,7 @@ app.get('/devtron', routes.devtron)
 app.get('/spectron', routes.spectron)
 app.use('/crowdin', routes.languages.proxy)
 
-app.get('/search', routes.search)
+app.use('/search', routes.search.index)
 
 // Generic 404 handler
 app.use(routes._404)
